+++ lib/autogram_sign.dart
@@ -3,4 +3,5 @@ library autogram_sign;
 export 'generated/autogram.swagger.dart';
 export 'src/autogram_authenticator.dart';
 export 'src/autogram_service.dart';
+export 'src/keys.dart';
 export 'src/service_exception.dart';